Fixes Mcollective Server Dependencies
Previously, the mcollective service would be in an "up" state
upon the successful deployment of a node.  Although, the service
would be "up", the broker would be unable to mco ping the
node.  This was due to the following reasons:

1. The service not being restarted after the associated config
files were changed by puppet.

2. DNS registration and resolution being configured after the
openshift_origin::mcollective_server class.  The node could not
resolve the broker's name, causing the mcollective service to
function inproperly until the service was restarted.