Ensure we don't attempt to change the node frontend
We want to ensure that the only time we change the node frontend is
on first puppet run. So create a marker file[1] via exec, create a file
based on an inline template that will update whenever there are changes
to the node_frontend_plugins. Diff those two before altering node.conf
or installing node plugins.