Refactor openshift_origin::register_dns to avoid using ensure_resource
We've discovered that puppetlabs-stdlib ensure_resource has a bug[1] that leads to the command parameter being ignored. This
seems to be addressed by upgrading to Puppet 3.3.0 or later, however we don't want to force users to do that.
Additionally, register_dns had a few issues, mostly centered around the fact that we were passing in a 'fqdn'
that really wasn't a FQDN but instead strings like 'register $role dns' which would then call an exec titled
'Register $fqdn'. Hopefully this simplifies things a bit.

1. https://tickets.puppetlabs.com/browse/MODULES-832